Reorganize main function.
